refactor: ProductStepper hook 분리
